requirements.txt
blinker==1.6.2
click==8.1.6
Flask==2.3.2
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
numpy==1.24.4
Pillow==10.0.0
scikit-learn==1.3.0
scipy==1.10.1
threadpoolctl==3.2.0
Werkzeug==2.3.6
zipp==3.16.2
